Split the Jar class into its base (zip) and extensions
A Jar is actually a Zip file with a manifest, but currently is a bit of
a mixture of both not clearly separated.

This splits the part that are more Zip-Specific into a base class 'Zip'
